add check before adding cpus to cpuset_reserved
There are some cases where None value is set to cpuset_reserved in
InstanceNUMATopology at _numa_fit_instance_cell() function in
hardware.py. However, libvirt driver treat cpuset_reserved value
as an iterate object when it constructs xml configuration.

To avoid a risk to get an error in libvirt driver, this patch adds
a check to see if the value is not None before adding the cpus
for emulator threads.

Change-Id: Iab3d950c4f4138118ac6a9fd98407eaadcb24d9e
Closes-Bug: #1746674