Use dict.get() when accessing capabilities dict
Many compute drivers override the driver capabilities dict [1]. The
code should be using dict.get() to avoid throwing KeyErrors if the
overriding dictionaries do not have the corresponding keys. If the
key is not found it will be assumed the the capability is not supported
and default to false. This ensures that no drivers are broken because
they are missing a capability.

[1] https://github.com/openstack/nova/blob/5251f18d87/nova/virt/driver.py#L124-L134

Change-Id: I1fa04fa110f2c65e10c065c61f2f0f58c1fad646
Closes-Bug: #1746608