Avoid raise InstanceNotRunning exception
when instance is being snapshot, it's possible that the instance
is deleted by another thread concurrently, libvirt driver translate
this InstanceNotFound exception into InstanceNotRunning and we didn't
catch and handle the exception and it leads to backtrace in compute log.

Also, fixes to catch the Exception and return a debug info
as the instance is anyway deleted already.

Change-Id: Ic6d0ac894fbbbf03fe967ee4f9ec93b2491cf073
Closes-Bug: 1759473