Bump compute RPC API to version 5.0
This moves us to version 5.0, with a proxy for 4.x for Queens. This includes
a lot of churn due to a lot of kwarg cruft. This is all just refactor, as it
doesn't include the client side change. So if this passes, we know we're
compatible, and subsequent patches will add client-side support.

Change-Id: I539b0a439030be57fd03012526ff0236f2b88a4e