Move the nova-next job in-tree and update it
This moves the legacy-tempest-dsvm-neutron-nova-next-full
job from openstack-zuul-jobs in-tree and updates a
few things since it's now branch specific:

- The cells v2 and placement stuff can be removed since those
  are required now (they were optional when this job would run
  against newton changes).
- Use nova/gate/post_test_hook.sh rather than the symlink under
  nova/tools/hooks/.

The job is currently non-voting and needs to remain that way
until bug 1747511 is fixed.

According to the branch restrictions in project-config, apparently
this job is only running against stable/newton (which no longer
exists) and master, presumably when master was pike. We should also
run this job against stable/pike for the service user test coverage
and post_test_hook.sh run, so we'll likely backport this change
to stable/pike as well.

The related project-config change is:

I36d96f89b3e5323746fcbcef5cc7e4d0384a184d

Change-Id: I24a5f73c29094a23e2fdef8ee8b43601300af593