Fix shelving a paused instance
It is possible to shelve a paused instance, but in that case the
guest is already shutdown, and some hypervisors will fail when
trying to perform a clean shutdown of a non-running guest. For
example, attempting to shelve a paused libvirt instance will
result in this error:

  libvirtError: Requested operation is not valid: domain is not running

Therefore, if the instance is paused, we don't attempt a clean
shutdown while shelving.

Related Tempest test: https://review.openstack.org/564127/

Closes-Bug: #1745529
Change-Id: I8ca25d9847d50001fbe8513a6c1dba8b697c24e4
(cherry picked from commit d5901f613cf98f61b5253a1568b22af1d9dd1a08)