Unmap compute nodes when deleting host mappings in delete cell operation
Presently when a host_mapping is deleted during the nova-manage cell_v2
delete_cell operation, the compute node record still remains with its
mapped field set to 1. Due to this, if we delete a cell, and later
create a new cell, the discover_hosts command will not work and the
compute node will not get mapped to the new cell. This patch fixes this
issue by setting it back to 0.

Change-Id: Iaf13f11bcd0bb25f673fc28f252cb9391409b9b5
Closes-Bug: #1747936