Don't JSON encode instance_info.traits for ironic
As of queens, the ironic virt driver pushes traits set on the flavor to
the ironic node's instance_info during instance spawn. This list of
traits is currently encoded as a JSON string, inside the JSON-encoded
instance_info. We should not use this double layer of JSON encoding, as
ironic expects instance_info.traits to be a list.

This change removes the nested JSON encoding when setting
node.instance_info.traits.

Change-Id: I1f3f96f7ef310022accde67cab1ae415bbc3c292
Closes-Bug: #1748942