Cleanup the manage-volumes admin doc
This was originaly copied over from "compute-manage-volumes.rst"
doc in the openstack-manuals repo but it's clearly mostly talking
about managing volumes on the Cinder side, not how they related
to compute server instances. Furthermore, Cinder already has a much
more detailed CLI guide on managing volumes, so this change removes
the content that's not specific to Nova and just links to the Cinder
docs.

Change-Id: I8e1c0d30debfdd4017c6d1a810a7b7dd382cd02d