Add request filter functionality to scheduler
This adds a pre-placement step where we can call a number of modular
"request filters" to modify the request_spec before we move on to
construct the call to placement to get allocation candidates.

No filters are added at this time, this just adds the infrastructure.

Related to blueprint placement-req-filter

Change-Id: I1535158a0dbd4a8527bb987e085e9391e5b0fde4