Add functional tests to ensure BDM removal on delete
In certain cases, such as when an instance fails to be scheduled,
the volume may already have an attachment created (or the volume
has been reserved in the old flows).

This patch adds a test to check that these volume attachments
are deleted and removed once the instance has been deleted. It
also adds some functionality to allow checking when an volume
has been reserved in the Cinder fixtures.

This backported patch drops the tests for the new-attach flow
for Cinder as it does not exist in Pike.

Change-Id: I85cc3998fbcde30eefa5429913ca287246d51255
Related-Bug: #1404867
(cherry picked from commit 20edeb362327ea9448d38b6e2bcb03e0a71060d0)