Store block device mappings in cell0
If an instance fails to get scheduled, it gets buried in cell0 but
none of it's block device mappings are stored.  At the API layer,
Nova reserves and creates attachments for new instances when
it gets a create request so these attachments are orphaned if the
block device mappings are not registered in the database somewhere.

This patch makes sure that if an instance is being buried in cell0,
all of it's block device mappings are recorded as well so they can
be later removed when the instance is deleted.

Conflicts:
      nova/conductor/manager.py

Change-Id: I64074923fb741fbf5459f66b8ab1a23c16f3303f
Related-Bug: #1404867
(cherry picked from commit ad9e2a568fcef4f69b185952352db9f651c37ad4)