Only attempt a rebuild claim for an evacuation to a new host
Change I11746d1ea996a0f18b7c54b4c9c21df58cc4714b changed the
behavior of the API and conductor when rebuilding an instance
with a new image such that the image is run through the scheduler
filters again to see if it will work on the existing host that
the instance is running on.

As a result, conductor started passing 'scheduled_node' to the
compute which was using it for logic to tell if a claim should be
attempted. We don't need to do a claim for a rebuild since we're
on the same host.

This removes the scheduled_node logic from the claim code, as we
should only ever attempt a claim if we're evacuating, which we
can determine based on the 'recreate' parameter.

Conflicts:
      nova/tests/functional/test_servers.py

NOTE(mriedem): test_rebuild_with_new_image does not exist in
Ocata and does not apply to Ocata since it is primarily
testing allocations getting created in Placement via the
FilterScheduler, which was new in Pike. As a result the change
to that test is not part of this backport, but a similar assertion
is added to an existing rebuild unit test.

Change-Id: I7fde8ce9dea16679e76b0cb2db1427aeeec0c222
Closes-Bug: #1750618
(cherry picked from commit a39029076c7997236a7f999682fb1e998c474204)
(cherry picked from commit 3c5e519a8875a9766a0d3a06cb76cceba26634e6)
(cherry picked from commit 9890f3f69622489a0fd57cb1df354d7aa60161e0)